feat(api): tag every Sentry alert with request origin, client IP, user-agent
Open a Sentry isolation scope around each HTTP request in
canonicalRequestLogging so every captureException / captureMessage fired
downstream — GraphQL errors, PostGraphile pool-connect failures,
log.error / Effect logError — inherits caller-identity tags automatically.

Adds searchable `origin` and `client_ip` tags to Sentry Issues plus a
`request` context entry with `user-agent`. Reuses the battle-tested
extractClientIp helper (x-real-ip preferred, rightmost x-forwarded-for
as fallback), moved from kg/searchInvocationLogger to utils/clientIp
now that it has more than one consumer.

Per-request withIsolationScope avoids tag bleed across concurrent
requests. No existing call sites need to be touched — scope-based tags
attach to all 19+ downstream capture sites automatically.
